WebApplication/requirements.txt
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
av==14.1.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
daphne==4.1.2
Django==5.1.4
django-cors-headers==4.6.0
django-redis==5.4.0
django-sslserver==0.22
django-sslserver==0.22
dotenv==0.9.9
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
fvcore==0.1.5.post20221221
huggingface-hub==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iopath==0.1.10
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
numpy==2.2.2
opencv-python==4.11.0.86
openncv==1.2
packaging==24.2
parameterized==0.9.0
pillow==11.1.0
portalocker==3.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pygame==2.6.1
pyOpenSSL==25.0.0
pyotp==2.9.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorchvideo==0.1.5
pywin32==308
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scipy==1.15.1
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
sympy==1.13.1
tabulate==0.9.0
termcolor==2.5.0
tokenizers==0.21.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.48.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
websocket-client==1.8.0
websockets==14.2
yacs==0.1.8
zope.interface==7.2
